made up responisive footer
