+++ pages/example.tsx
@@ -1,0 +1,9 @@
+import type { NextPage } from "next";
+import Footer from "../components/footer/Footer";
+
+
+const Example = () => (
+    <Footer/>
+);
+
+export default Example;
